style: Extract CoinCard into reusable component and replace inline coin card rendering in App
